+++ src/mastodon_api/instance/views.rs
@@ -3,8 +3,8 @@ use actix_web::{get, web, HttpResponse, Scope};
 use mitra_config::Config;
 
 use crate::database::{get_database_client, DbPool};
-use crate::errors::HttpError;
 use crate::ethereum::contracts::ContractSet;
+use crate::mastodon_api::errors::MastodonError;
 use crate::models::{
     instances::queries::get_peer_count,
     posts::queries::get_local_post_count,
@@ -18,7 +18,7 @@ async fn instance_view(
     config: web::Data<Config>,
     db_pool: web::Data<DbPool>,
     maybe_blockchain: web::Data<Option<ContractSet>>,
-) -> Result<HttpResponse, HttpError> {
+) -> Result<HttpResponse, MastodonError> {
     let db_client = &**get_database_client(&db_pool).await?;
     let user_count = get_user_count(db_client).await?;
     let post_count = get_local_post_count(db_client).await?;