+++ src/mastodon_api/notifications/views.rs
@@ -11,9 +11,9 @@ use actix_web_httpauth::extractors::bearer::BearerAuth;
 use mitra_config::Config;
 
 use crate::database::{get_database_client, DbPool};
-use crate::errors::HttpError;
 use crate::http::get_request_base_url;
 use crate::mastodon_api::{
+    errors::MastodonError,
     oauth::auth::get_current_user,
     pagination::get_paginated_response,
 };
@@ -28,7 +28,7 @@ async fn get_notifications_view(
     db_pool: web::Data<DbPool>,
     query_params: web::Query<NotificationQueryParams>,
     request: HttpRequest,
-) -> Result<HttpResponse, HttpError> {
+) -> Result<HttpResponse, MastodonError> {
     let db_client = &**get_database_client(&db_pool).await?;
     let current_user = get_current_user(db_client, auth.token()).await?;
     let base_url = get_request_base_url(connection_info);